Refaktorer saksgangtogler til egen komponent
